# shared/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Data Voyager API
  version: 0.1.0
  description: Data Voyager REST API
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /settings/ai:
    get:
      operationId: getAISettings
      summary: Get current AI settings (no secret values)
      tags: [settings]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AISettingsData"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      operationId: updateAISettings
      summary: Save AI settings (empty api_key keeps existing value)
      tags: [settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAISettingsRequest"
      responses:
        "204":
          description: No Content
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  schemas:
...
    ClaudeSettingsResponse:
      type: object
      properties:
        api_key_set:
          type: boolean
        model:
          type: string
        base_url:
          type: string
    OpenAISettingsResponse:
      type: object
      properties:
        api_key_set:
          type: boolean
        model:
          type: string
        base_url:
          type: string
    CopilotSettingsResponse:
      type: object
      properties:
        api_key_set:
          type: boolean
        model:
          type: string
        base_url:
          type: string
    OllamaSettingsResponse:
      type: object
      properties:
        model:
          type: string
        base_url:
          type: string
    AISettingsData:
      type: object
      required: [enabled, provider]
      properties:
        enabled:
          type: boolean
        provider:
          type: string
        claude:
          $ref: "#/components/schemas/ClaudeSettingsResponse"
        openai:
          $ref: "#/components/schemas/OpenAISettingsResponse"
        copilot:
          $ref: "#/components/schemas/CopilotSettingsResponse"
        ollama:
          $ref: "#/components/schemas/OllamaSettingsResponse"
...
    ClaudeSettingsInput:
      type: object
      properties:
        api_key:
          type: string
          description: Empty string keeps the existing key
        model:
          type: string
        base_url:
          type: string
    OpenAISettingsInput:
      type: object
      properties:
        api_key:
          type: string
          description: Empty string keeps the existing key
        model:
          type: string
        base_url:
          type: string
    CopilotSettingsInput:
      type: object
      properties:
        api_key:
          type: string
          description: Empty string keeps the existing key
        model:
          type: string
        base_url:
          type: string
    OllamaSettingsInput:
      type: object
      properties:
        model:
          type: string
        base_url:
          type: string
    UpdateAISettingsRequest:
      type: object
      properties:
        enabled:
          type: boolean
        provider:
          type: string
        claude:
          $ref: "#/components/schemas/ClaudeSettingsInput"
        openai:
          $ref: "#/components/schemas/OpenAISettingsInput"
        copilot:
          $ref: "#/components/schemas/CopilotSettingsInput"
        ollama:
          $ref: "#/components/schemas/OllamaSettingsInput"
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...